+++ exonerations.py
@@ -20,10 +20,10 @@ def median_age(data):
 
     print('Median age at time of arrest: %i' % median_age)
 
-    data['exonerations'].bins('age', 10, 0, 100).print_bars(width=80)
+    data['exonerations'].bins('age', 10, 0, 100).print_bars('age', width=80)
     data['exonerations'].pivot('age').order_by('age').print_bars('age', width=80)
 
-    data['exonerations'].bins('age').print_bars(width=80)
+    data['exonerations'].bins('age').print_bars('age', width=80)
 
 
 def years_in_prison(data):